+++ CHANGES.md
@@ -4,7 +4,10 @@
 
 - Fix Docker compose file, so example data can be loaded into database (author @zstatmanweil, https://github.com/stac-utils/stac-fastapi-pgstac/pull/142)
 - Handle `next` and `dev` tokens now returned as links from pgstac>=0.9.0 (author @zstatmanweil, https://github.com/stac-utils/stac-fastapi-pgstac/pull/140)
+- Add collection search extension ([#139](https://github.com/stac-utils/stac-fastapi-pgstac/pull/139))
 - Fix `filter` extension implementation in `CoreCrudClient`
+- update `pypgstac` requirement to `>=0.8,<0.10`
+- set `pypgstac==0.9.*` for test requirements
 
 ## [3.0.0] - 2024-08-02